Organize tests a bit
