+++ test/models/sku.rb
@@ -1,0 +1,3 @@
+class Sku
+  searchkick callbacks: defined?(ActiveJob) ? :async : true
+end
