+++ test/models/song.rb
@@ -1,0 +1,7 @@
+class Song
+  searchkick
+
+  def search_routing
+    name
+  end
+end
